// src/shared/zero/schema.gen.ts
const eloMatchupsTable = {
  name: "eloMatchups",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    recordAId: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "record_a_id",
    },
    recordBId: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "record_b_id",
    },
    winnerId: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "winner_id",
    },
    recordType: {
      type: "string",
      optional: false,
      customType: null as unknown as "entity" | "concept" | "artifact",
      serverName: "record_type",
    },
    recordCreatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
  },
  primaryKey: ["id"],
  serverName: "elo_matchups",
} as const;
const linksTable = {
  name: "links",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    sourceId: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "source_id",
    },
    targetId: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "target_id",
    },
    predicate: {
      type: "string",
      optional: false,
      customType: null as unknown as
        | "created_by"
        | "creator_of"
        | "via"
        | "source_for"
        | "edited_by"
        | "editor_of"
        | "translated_by"
        | "translator_of"
        | "contained_by"
        | "contains"
        | "quotes"
        | "quoted_in"
        | "has_format"
        | "format_of"
        | "tagged_with"
        | "tag_of"
        | "references"
        | "referenced_by"
        | "about"
        | "subject_of"
        | "responds_to"
        | "responded_by"
        | "related_to"
        | "counters"
        | "countered_by"
        | "same_as",
    },
    notes: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    recordCreatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
    recordUpdatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "updated_at",
    },
  },
  primaryKey: ["id"],
} as const;
const mediaTable = {
  name: "media",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    recordId: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "record_id",
    },
    url: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    altText: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "alt_text",
    },
    altTextGeneratedAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "alt_text_generated_at",
    },
    type: {
      type: "string",
      optional: false,
      customType: null as unknown as
        | "message"
        | "application"
        | "audio"
        | "font"
        | "image"
        | "model"
        | "multipart"
        | "text"
        | "video",
    },
    format: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
    },
    contentTypeString: {
      type: "string",
      optional: false,
      customType: null as unknown as string,
      serverName: "content_type_string",
    },
    fileSize: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "file_size",
    },
    width: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
    height: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
    },
    versionOfMediaId: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "version_of_media_id",
    },
    recordCreatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
    recordUpdatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "updated_at",
    },
  },
  primaryKey: ["id"],
} as const;
const recordsTable = {
  name: "records",
  columns: {
    id: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
    },
    slug: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    type: {
      type: "string",
      optional: false,
      customType: null as unknown as "entity" | "concept" | "artifact",
    },
    title: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    sense: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    abbreviation: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    url: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    avatarUrl: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "avatar_url",
    },
    summary: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    content: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    notes: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
    },
    mediaCaption: {
      type: "string",
      optional: true,
      customType: null as unknown as string,
      serverName: "media_caption",
    },
    eloScore: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "elo_score",
    },
    isPrivate: {
      type: "boolean",
      optional: false,
      customType: null as unknown as boolean,
      serverName: "is_private",
    },
    isCurated: {
      type: "boolean",
      optional: false,
      customType: null as unknown as boolean,
      serverName: "is_curated",
    },
    reminderAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "reminder_at",
    },
    sources: {
      type: "json",
      optional: true,
      customType: null as unknown as (
        | "ai_chat"
        | "airtable"
        | "browser_history"
        | "crawler"
        | "embeddings"
        | "feedbin"
        | "github"
        | "lightroom"
        | "manual"
        | "raindrop"
        | "readwise"
        | "twitter"
      )[],
    },
    recordCreatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "created_at",
    },
    recordUpdatedAt: {
      type: "number",
      optional: false,
      customType: null as unknown as number,
      serverName: "updated_at",
    },
    contentCreatedAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "content_created_at",
    },
    contentUpdatedAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "content_updated_at",
    },
    textEmbeddedAt: {
      type: "number",
      optional: true,
      customType: null as unknown as number,
      serverName: "text_embedded_at",
    },
  },
  primaryKey: ["id"],
} as const;
const eloMatchupsRelationships = {
  recordA: [
    {
      sourceField: ["recordAId"],
      destField: ["id"],
      destSchema: "records",
      cardinality: "one",
    },
  ],
  recordB: [
    {
      sourceField: ["recordBId"],
      destField: ["id"],
      destSchema: "records",
      cardinality: "one",
    },
  ],
  winner: [
    {
      sourceField: ["winnerId"],
      destField: ["id"],
      destSchema: "records",
      cardinality: "one",
    },
  ],
} as const;
const mediaRelationships = {
  versionOf: [
    {
      sourceField: ["versionOfMediaId"],
      destField: ["id"],
      destSchema: "media",
      cardinality: "one",
    },
  ],
  versions: [
    {
      sourceField: ["id"],
      destField: ["versionOfMediaId"],
      destSchema: "media",
      cardinality: "many",
    },
  ],
  record: [
    {
      sourceField: ["recordId"],
      destField: ["id"],
      destSchema: "records",
      cardinality: "one",
    },
  ],
} as const;
const linksRelationships = {
  source: [
    {
      sourceField: ["sourceId"],
      destField: ["id"],
      destSchema: "records",
      cardinality: "one",
    },
  ],
  target: [
    {
      sourceField: ["targetId"],
      destField: ["id"],
      destSchema: "records",
      cardinality: "one",
    },
  ],
} as const;
const recordsRelationships = {
  outgoingLinks: [
    {
      sourceField: ["id"],
      destField: ["sourceId"],
      destSchema: "links",
      cardinality: "many",
    },
  ],
  incomingLinks: [
    {
      sourceField: ["id"],
      destField: ["targetId"],
      destSchema: "links",
      cardinality: "many",
    },
  ],
  eloMatchupsAsA: [
    {
      sourceField: ["id"],
      destField: ["recordAId"],
      destSchema: "eloMatchups",
      cardinality: "many",
    },
  ],
  eloMatchupsAsB: [
    {
      sourceField: ["id"],
      destField: ["recordBId"],
      destSchema: "eloMatchups",
      cardinality: "many",
    },
  ],
  media: [
    {
      sourceField: ["id"],
      destField: ["recordId"],
      destSchema: "media",
      cardinality: "many",
    },
  ],
} as const;
/**
 * The Zero schema object.
 * This type is auto-generated from your Drizzle schema definition.
 */
export const schema = {
  tables: {
    eloMatchups: eloMatchupsTable,
    links: linksTable,
    media: mediaTable,
    records: recordsTable,
  },
  relationships: {
    eloMatchups: eloMatchupsRelationships,
    media: mediaRelationships,
    links: linksRelationships,
    records: recordsRelationships,
  },
  enableLegacyQueries: false,
  enableLegacyMutators: false,
} as const;
// ...
export const zql = createBuilder(schema);